+++ compiler/src/main/java/info/teksol/mc/messages/ListMessageLogger.java
@@ -1,6 +1,5 @@
 package info.teksol.mc.messages;
 
-import info.teksol.mc.emulator.EmulatorMessage;
 import org.jspecify.annotations.NullMarked;
 
 import java.util.ArrayList;
@@ -10,8 +9,6 @@ import java.util.List;
 public class ListMessageLogger extends AbstractMessageLogger {
     private final MessageConsumer messageConsumer;
     protected final List<MindcodeMessage> messages = new ArrayList<>();
-    private boolean hasErrors = false;
-    private boolean hasWarnings = false;
 
     public ListMessageLogger() {
         this.messageConsumer = s -> {};
@@ -23,29 +20,11 @@ public class ListMessageLogger extends AbstractMessageLogger {
 
     @Override
     public void addMessage(MindcodeMessage message) {
-        if (!(message instanceof EmulatorMessage)) {
-            switch (message.level()) {
-                case ERROR -> hasErrors = true;
-                case WARNING -> hasWarnings = true;
-            }
-        }
         messages.add(message);
         messageConsumer.accept(message);
     }
 
     public List<MindcodeMessage> getMessages() {
         return messages;
-    }
-
-    public boolean hasErrors() {
-        return hasErrors;
-    }
-
-    public boolean hasWarnings() {
-        return hasWarnings;
-    }
-
-    public boolean hasErrorsOrWarnings() {
-        return hasErrors || hasWarnings;
     }
 }