+++ exttest/src/main/java/info/teksol/mindcode/exttest/cases/BasicTestCaseExecutor.java
@@ -40,7 +40,7 @@ public class BasicTestCaseExecutor implements TestCaseExecutor {
                     .map(Assertion::generateErrorMessage)
                     .collect(Collectors.joining("\n"));
 
-            boolean success = unexpectedMessages.isEmpty() && failedTests.isEmpty() && !compiler.isRuntimeError();
+            boolean success = unexpectedMessages.isEmpty() && failedTests.isEmpty() && !compiler.hasRuntimeError();
 
             if (success) {
                 if (compiler.globalCompilerProfile().isRun() && compiler.getAssertions().isEmpty()) {